move all unit tests file in one folder
