+++ test/unit/differ/data/different-unnoticeable.png
