+++ test/unit/differ/data/different.png
