+++ test/unit/differ/data/same.png
