+++ test/unit/differ/data/strict-unnoticeable.png
