+++ test/unit/differ/data/strict.png
