+++ test/unit/differ/data/tall.png
