+++ test/unit/differ/data/wide.png
