+++ test/unit/differ/look-same.js
@@ -1,5 +1,5 @@
 var expect = require('chai').expect;
-var LooksSameAdapter = require('../../lib/adapters/looks-same.js');
+var LooksSameAdapter = require('../../../lib/adapters/looks-same.js');
 var fs = require('fs');
 var path = require('path');
